# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and account-security API: registration with email verification, JWT login with TOTP MFA, Google OAuth 2.0, session and device management, and passkeys/WebAuthn. The response envelope is {code, message, data} on every endpoint.",
        "title": "FinnApiGo Auth API",
        "contact": {
            "name": "FinnApiGo",
            "url": "https://github.com/NguyenQuan121321/FinnApiGo/issues"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists the caller's active (non-expired, non-revoked) sessions/devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "List active sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.SessionsEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes a single session by ID, scoped to the caller. The target device can no longer rotate its refresh token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sessions"
                ],
                "summary": "Revoke a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.NullDataEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/trusted-devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists caller's active trusted devices eligible for 30-day MFA bypass.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TrustedDevices"
                ],
                "summary": "List trusted devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TrustedDevicesListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/trusted-devices/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes MFA bypass trust for a specific device.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TrustedDevices"
                ],
                "summary": "Revoke trusted device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.NullDataEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "swagger.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "validation error"
                }
            }
        },
...
        "swagger.NullDataEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
...
        "swagger.SessionInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-01-15T10:30:00Z"
                },
                "deviceName": {
                    "type": "string",
                    "example": "Chrome on Windows"
                },
                "expiresAt": {
                    "type": "string",
                    "example": "2026-01-22T10:30:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
                },
                "ipAddress": {
                    "type": "string",
                    "example": "192.168.1.100"
                },
                "isCurrent": {
                    "type": "boolean",
                    "example": true
                },
                "lastActiveAt": {
                    "type": "string",
                    "example": "2026-01-15T11:00:00Z"
                },
                "locationEstimate": {
                    "type": "string",
                    "example": "Ho Chi Minh City, VN"
                },
                "userAgent": {
                    "type": "string",
                    "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)"
                }
            }
        },
        "swagger.SessionsData": {
            "type": "object",
            "properties": {
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.SessionInfo"
                    }
                }
            }
        },
        "swagger.SessionsEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.SessionsData"
                },
                "message": {
                    "type": "string",
                    "example": "sessions fetched"
                }
            }
        },
...
        "swagger.TrustedDeviceInfoData": {
            "type": "object",
            "properties": {
                "deviceName": {
                    "type": "string",
                    "example": "Chrome MacBook Pro"
                },
                "expiresAt": {
                    "type": "string",
                    "example": "2026-10-03T18:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "ipAddress": {
                    "type": "string",
                    "example": "192.168.1.50"
                },
                "lastUsedAt": {
                    "type": "string",
                    "example": "2026-09-03T18:00:00Z"
                }
            }
        },
        "swagger.TrustedDevicesListData": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.TrustedDeviceInfoData"
                    }
                }
            }
        },
        "swagger.TrustedDevicesListEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.TrustedDevicesListData"
                },
                "message": {
                    "type": "string",
                    "example": "trusted devices retrieved"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Send as \"Bearer \u003ctoken\u003e\". For /auth/mfa/login-verify this must be the short-lived mfa_pending token issued by /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SudoAuth": {
            "description": "Short-lived sudo token minted by POST /auth/mfa/totp/recovery-codes; required for privileged operations.",
            "type": "apiKey",
            "name": "X-Sudo-Token",
            "in": "header"
        }
    }
}